feat: 1229 setup backend route for getting listings
