+++ packages/backend/src/services/auth-service/adapters/msal.ts
@@ -162,7 +162,7 @@ const acquireToken = (options: AuthOptions = defaultOptions) => {
           postLogoutRedirectUri: options.postLogoutRedirectUri,
         })(ctx)
       }
-
+      console.log("error: ", error)
       ctx.next(error)
     }
   }
@@ -207,6 +207,7 @@ const handleRedirect = (options: AuthOptions = defaultOptions) => {
 
       return ctx.redirect(state.successRedirect)
     } catch (error) {
+      console.log("error: ", error)
       ctx.next(error)
     }
   }
@@ -296,6 +297,7 @@ const redirectToAuthCodeUrl = (
       )
       ctx.redirect(authCodeUrlResponse)
     } catch (error) {
+      console.log("error", error)
       ctx.next(error)
     }
   }
